% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brxx_general.R
\name{brxx_general}
\alias{brxx_general}
\title{brxx_general: Bayesian Estimation of Reliability from Variance Estimates}
\usage{
brxx_general(a, b, alpha, beta, CI)
}
\arguments{
\item{a}{True score variance estimate.}

\item{b}{Error variance estimate.}

\item{alpha}{Prior true score variance.}

\item{beta}{Prior error variance.}

\item{CI}{Credible interval quantile, as a decimal (ie, for 95 percent, 0.95).}
}
\value{
Returns estimated median and quantile based credible limits for reliability.
}
\description{
This function estimates reliability from given true and error variance estimates.
}
\examples{
a=18.7
b=3.3
alpha=3.51
beta=1.75
brxx_general(a=a,b=b,alpha=alpha,beta=beta,CI=0.95)

}
